+++ src/pages/registration/RegistrationForm.tsx
@@ -20,7 +20,7 @@ import { RootState } from '../../redux/store';
 import { RegistrationFormLocationState } from '../../types/locationState.types';
 import { Registration, RegistrationStatus, RegistrationTab } from '../../types/registration.types';
 import { getTouchedTabFields, validateRegistrationForm } from '../../utils/formik-helpers/formik-helpers';
-import { getTitleString } from '../../utils/registration-helpers';
+import { getTitleString, userHasAccessRight } from '../../utils/registration-helpers';
 import { createUppy } from '../../utils/uppy/uppy-config';
 import { doNotRedirectQueryParam, UrlPathTemplate } from '../../utils/urlPaths';
 import { Forbidden } from '../errorpages/Forbidden';
@@ -64,7 +64,7 @@ export const RegistrationForm = ({ identifier }: RegistrationFormProps) => {
   const initialTabNumber = new URLSearchParams(location.search).get('tab');
   const [tabNumber, setTabNumber] = useState(initialTabNumber ? +initialTabNumber : RegistrationTab.Description);
 
-  const canEditRegistration = true; // userHasAccessRight(registration, 'update');
+  const canEditRegistration = userHasAccessRight(registration, 'update');
 
   useEffect(() => {
     window.scrollTo({ top: 0, behavior: 'smooth' });
